// apps/web/lib/transport/http/http-error.types.ts
export enum HttpErrorName {
  Unauthorized = 'UNAUTHORIZED',
  BadRequest = 'BAD_REQUEST',
  ValidationError = 'VALIDATION_ERROR',
  InternalServerError = 'INTERNAL_SERVER_ERROR',
  Forbidden = 'FORBIDDEN',
  NotFound = 'NOT_FOUND',
  TooManyRequests = 'TOO_MANY_REQUESTS',
  ServiceUnavailable = 'SERVICE_UNAVAILABLE',
  GatewayTimeout = 'GATEWAY_TIMEOUT',
  BadGateway = 'BAD_GATEWAY',
  RequestTimeout = 'REQUEST_TIMEOUT'
}

export type HttpError = {
  name: HttpErrorName
  status: number
  message: string
  retryable: boolean
}